requirements.txt
pydantic
fastapi
uvicorn
PyYAML
typing-extensions>=3.10